+++ src/es/upm/dit/tfg/webLab/servlets/BorrarPlazaServlet.java
@@ -1,6 +1,8 @@
 package es.upm.dit.tfg.webLab.servlets;
 
 import org.apache.log4j.Logger;
+import org.apache.shiro.subject.Subject;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -22,25 +24,36 @@ public class BorrarPlazaServlet extends HttpServlet{
 	private final static Logger log = Logger.getLogger(BorrarPlazaServlet.class);
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException, java.io.IOException {
+		Subject currentUser = (Subject) req.getSession().getAttribute("currentUser");
 		req.getSession().removeAttribute("mensaje");
 		int id = Integer.parseInt(req.getParameter("id"));
 
-		Plaza plaza = PlazaDAOImplementation.getInstance().readPlaza(id);
-		PlazaDAOImplementation.getInstance().deletePlaza(plaza);
 		
-		Usuario usuario = (Usuario) req.getSession().getAttribute("usuario");
-
-		//log.info("El usuario "+usuario.getNombre()+" "+usuario.getApellidos()+" ha borrado la plaza de profesor "+plaza.getPlaza());
-		
-		
-		List<Plaza> todasPlazas = PlazaDAOImplementation.getInstance().readPlazas();
+		/*
+		 * Solo puede entrar aquí si es administrador o si tiene el rol para gestionar usuarios 
+		 */
+		if (currentUser.hasRole("administrador") || currentUser.hasRole("gestionusuarios")){
+			Plaza plaza = PlazaDAOImplementation.getInstance().readPlaza(id);
+			PlazaDAOImplementation.getInstance().deletePlaza(plaza);
+			
+			Usuario usuario = (Usuario) req.getSession().getAttribute("usuario");
 	
+			//log.info("El usuario "+usuario.getNombre()+" "+usuario.getApellidos()+" ha borrado la plaza de profesor "+plaza.getPlaza());
+			
+			
+			List<Plaza> todasPlazas = PlazaDAOImplementation.getInstance().readPlazas();
 		
-		req.getSession().setAttribute("plazas", todasPlazas);
-		
+			
+			req.getSession().setAttribute("plazas", todasPlazas);
+			
+	
+			String msj = "Plaza borrada con éxito";
+			req.getSession().setAttribute("mensaje", msj);
 
-		String msj = "Plaza borrada con éxito";
-		req.getSession().setAttribute("mensaje", msj);
-		resp.sendRedirect(req.getContextPath()+ "/CRUDPlaza.jsp");
+			getServletContext().getRequestDispatcher("/CRUDPlaza.jsp").forward(req, resp);
+			
+		}else {
+			getServletContext().getRequestDispatcher("/NoPermitido.jsp").forward(req, resp);
+		}
 	}
 }
