+++ src/es/upm/dit/tfg/webLab/servlets/BorrarProfesorServlet.java
@@ -1,6 +1,8 @@
 package es.upm.dit.tfg.webLab.servlets;
 
 import org.apache.log4j.Logger;
+import org.apache.shiro.subject.Subject;
+
 import java.util.List;
 
 import javax.servlet.ServletException;
@@ -22,26 +24,36 @@ public class BorrarProfesorServlet extends HttpServlet{
 	private final static Logger log = Logger.getLogger(BorrarProfesorServlet.class);
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException, java.io.IOException {
+		Subject currentUser = (Subject) req.getSession().getAttribute("currentUser");
 		req.getSession().removeAttribute("mensaje");
 		int idProfe = Integer.parseInt(req.getParameter("idProfe"));
 		int idUsuario = Integer.parseInt(req.getParameter("idUsuario"));
 
-		Profesor profesor = ProfesorDAOImplementation.getInstance().readProfesor(idProfe);
-		ProfesorDAOImplementation.getInstance().deleteProfesor(profesor);
 		
-		Usuario usuario = UsuarioDAOImplementation.getInstance().readUsuario(idUsuario);
-		UsuarioDAOImplementation.getInstance().deleteUsuario(usuario);
-		
-		Usuario usuarioAccion = (Usuario) req.getSession().getAttribute("usuario");
+		/*
+		 * Solo puede entrar aquí si es administrador o si tiene el rol para gestionar usuarios 
+		 */
+		if (currentUser.hasRole("administrador") || currentUser.hasRole("gestionusuarios")){
+			Profesor profesor = ProfesorDAOImplementation.getInstance().readProfesor(idProfe);
+			ProfesorDAOImplementation.getInstance().deleteProfesor(profesor);
+			
+			Usuario usuario = UsuarioDAOImplementation.getInstance().readUsuario(idUsuario);
+			UsuarioDAOImplementation.getInstance().deleteUsuario(usuario);
+			
+			Usuario usuarioAccion = (Usuario) req.getSession().getAttribute("usuario");
+	
+			//log.info("El usuario "+usuarioAccion.getNombre()+" "+usuarioAccion.getApellidos()+" ha borrado al profesor "+usuario.getNombre()+" "+usuario.getApellidos());
+	
+			
+			List<Profesor> todasProfesores = ProfesorDAOImplementation.getInstance().readProfesores();
+			req.getSession().setAttribute("profesores", todasProfesores);
+			
+			String msj = "Profesor borrado con éxito";
+			req.getSession().setAttribute("mensaje", msj);
 
-		//log.info("El usuario "+usuarioAccion.getNombre()+" "+usuarioAccion.getApellidos()+" ha borrado al profesor "+usuario.getNombre()+" "+usuario.getApellidos());
-
-		
-		List<Profesor> todasProfesores = ProfesorDAOImplementation.getInstance().readProfesores();
-		req.getSession().setAttribute("profesores", todasProfesores);
-		
-		String msj = "Profesor borrado con éxito";
-		req.getSession().setAttribute("mensaje", msj);
-		resp.sendRedirect(req.getContextPath()+ "/CRUDProfesor.jsp");
+			getServletContext().getRequestDispatcher("/CRUDProfesor.jsp").forward(req, resp);
+		}else {
+			getServletContext().getRequestDispatcher("/NoPermitido.jsp").forward(req, resp);
+		}
 	}
 }
