+++ src/es/upm/dit/tfg/webLab/servlets/BorrarUsuarioServlet.java
@@ -1,6 +1,8 @@
 package es.upm.dit.tfg.webLab.servlets;
 
 import org.apache.log4j.Logger;
+import org.apache.shiro.subject.Subject;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -21,30 +23,41 @@ public class BorrarUsuarioServlet extends HttpServlet{
 	private final static Logger log = Logger.getLogger(BorrarUsuarioServlet.class);
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException, java.io.IOException {
+		Subject currentUser = (Subject) req.getSession().getAttribute("currentUser");
 		req.getSession().removeAttribute("mensaje");
 		int id = Integer.parseInt(req.getParameter("id"));
 		
-		Usuario usuario = UsuarioDAOImplementation.getInstance().readUsuario(id);
-		UsuarioDAOImplementation.getInstance().deleteUsuario(usuario);
 		
-		Usuario usuarioAccion = (Usuario) req.getSession().getAttribute("usuario");
+		/*
+		 * Solo puede entrar aquí si es administrador o si tiene el rol para gestionar usuarios 
+		 */
+		if (currentUser.hasRole("administrador") || currentUser.hasRole("gestionusuarios")){
+			Usuario usuario = UsuarioDAOImplementation.getInstance().readUsuario(id);
+			UsuarioDAOImplementation.getInstance().deleteUsuario(usuario);
+			
+			Usuario usuarioAccion = (Usuario) req.getSession().getAttribute("usuario");
+	
+			//log.info("El usuario "+usuarioAccion.getNombre()+" "+usuarioAccion.getApellidos()+" ha borrado el usuario "+usuario.getNombre()+" "+usuario.getApellidos());
+	
+			
+			List<Usuario> todoUsuarios = UsuarioDAOImplementation.getInstance().readUsuarios();
+			List<Usuario> usuarios =new ArrayList();
+			
+			for (int i=0; i<todoUsuarios.size(); i++) {
+				Profesor profesor = todoUsuarios.get(i).getProfesor();
+				if(profesor==null) usuarios.add(todoUsuarios.get(i));
+			}
+			
+			req.getSession().setAttribute("usuarios", usuarios);
+	
+			
+			String msj = "Usuario editado con éxito";
+			req.getSession().setAttribute("mensaje", msj);
 
-		//log.info("El usuario "+usuarioAccion.getNombre()+" "+usuarioAccion.getApellidos()+" ha borrado el usuario "+usuario.getNombre()+" "+usuario.getApellidos());
-
-		
-		List<Usuario> todoUsuarios = UsuarioDAOImplementation.getInstance().readUsuarios();
-		List<Usuario> usuarios =new ArrayList();
-		
-		for (int i=0; i<todoUsuarios.size(); i++) {
-			Profesor profesor = todoUsuarios.get(i).getProfesor();
-			if(profesor==null) usuarios.add(todoUsuarios.get(i));
+			getServletContext().getRequestDispatcher("/CRUDPAS.jsp").forward(req, resp);
+			
+		}else {
+			getServletContext().getRequestDispatcher("/NoPermitido.jsp").forward(req, resp);
 		}
-		
-		req.getSession().setAttribute("usuarios", usuarios);
-
-		
-		String msj = "Usuario editado con éxito";
-		req.getSession().setAttribute("mensaje", msj);
-		resp.sendRedirect(req.getContextPath() + "/CRUDPAS.jsp");
 	}
 }
